mcp_server/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-dotenv
openai==1.30.1
pydantic==2.7.1
simpleeval==0.9.13 
google-generativeai==0.5.4